% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prenoms.R
\docType{data}
\name{departements}
\alias{departements}
\title{Departements}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 101 rows and 5 columns.}
\source{
https://www.data.gouv.fr/fr/datasets/contours-des-departements-francais-issus-d-openstreetmap/
}
\usage{
departements
}
\description{
Departements
}
\keyword{datasets}
