% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prenoms}
\alias{prenoms}
\title{Prenoms des garcons et filles du Quebec de 1980 a 2017}
\format{Une base de donnees avec 13 511 308 lignes et 5 colonnes:
\describe{
  \item{annee}{Annee}
  \item{sexe}{F pour feminin et M pour masculin}
  \item{prenom}{Prenom}
  \item{n}{Frequence d apparition du prenom}
  \item{prop}{Proportion du prenom}
}}
\source{
\url{https://www.donneesquebec.ca/recherche/fr/dataset/banque-de-prenoms-garcons}

\url{https://www.donneesquebec.ca/recherche/fr/dataset/banque-de-prenoms-filles}
}
\usage{
prenoms
}
\description{
Une base de donnees contenant les prenoms masculins et feminins au
Quebec de 1980 a 2017.
}
\keyword{datasets}
