% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prenoms}
\alias{prenoms}
\title{Prenoms des bebes du Quebec entre 1980 et 2017}
\format{Une base de donnees contenant 13 511 308 lignes et 5 colonnes:
\describe{
  \item{annee}{Annee}
  \item{sexe}{F pour feminin et M pour masculin}
  \item{prenom}{Prenom}
  \item{n}{Frequence}
  \item{prop}{Proportion}
}}
\source{
\url{https://www.donneesquebec.ca/recherche/fr/dataset/banque-de-prenoms-garcons}

\url{https://www.donneesquebec.ca/recherche/fr/dataset/banque-de-prenoms-filles}
}
\usage{
prenoms
}
\description{
Une base de donnes contenant les prenoms des enfants du Quebec
entre 1980 et 2017.
}
\keyword{datasets}
