% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prenoms}
\alias{prenoms}
\title{Prenoms des bebes du Quebec entre 1980 et 2019}
\format{
Une base de donnees contenant 744 662 lignes et 5 colonnes:
\describe{
  \item{annee}{Annee}
  \item{sexe}{F pour feminin et M pour masculin}
  \item{prenom}{Prenom}
  \item{n}{Frequence}
  \item{prop}{Proportion}
}
}
\source{
\url{https://www.donneesquebec.ca/recherche/fr/dataset/banque-de-prenoms-garcons}

\url{https://www.donneesquebec.ca/recherche/fr/dataset/banque-de-prenoms-filles}
}
\usage{
prenoms
}
\description{
Une base de donnes contenant les prenoms des enfants du Quebec
entre 1980 et 2019.
}
\examples{
\dontrun{
  library(dplyr)
  library(ggplot2)
  library(prenoms)

  # Prenoms des membres de ma famille
  famille <- prenoms \%>\%
    filter(
      prenom == "Marc-Andre" & sexe == "M" |
      prenom == "Laurent" & sexe == "M" |
      prenom == "Melanie" & sexe == "F" |
      prenom == "Anna" & sexe == "F"
    ) \%>\%
    group_by(prenom, annee, sexe) \%>\%
    summarise(n = sum(n)) \%>\%
    arrange(annee)

  # Representation ggplot
  ggplot(data = famille, aes(x = annee, y = n, color = prenom))+
    geom_line()+
    scale_x_continuous( breaks = seq(1980, 2020, by = 5))
}
}
\keyword{datasets}
